# utils/Swagger/oas.yaml
openapi: 3.0.3
servers:
  - url: 'https://test.es.sign.fiskaly.com/api/v1'
    description: |
      Entorno de prueba (Test)
  - url: 'https://live.es.sign.fiskaly.com/api/v1'
    description: |
      Entorno de producción (Live)
info:
  version: '1.13.0'
  title: fiskaly SIGN ES
  contact:
    email: 'info@fiskaly.com'
    name: 'Contacto'
    url: 'https://www.fiskaly.com/es'
  description: |
    Información Legal: [fiskaly.com/impressum](https://www.fiskaly.com/es/aviso-legal)
    <br>
    Política de Privacidad: [fiskaly.com/datenschutz](https://www.fiskaly.com/es/privacy-policy)
    <br>
    Solución: [fiskaly.com/signes](https://www.fiskaly.com/es/signes)
    <br>
    Guía para Desarrolladores: [developer.fiskaly.com/es/sign-es](https://developer.fiskaly.com/es/sign-es/introduccion)

    # Introducción

    **SIGN ES**, la [solución](https://www.fiskaly.com/es/signes) de fiskaly, ofrece una [RESTful](https://es.wikipedia.org/wiki/Transferencia_de_Estado_Representacional) API que asegura el cumplimiento con las normativas fiscales españolas como **TicketBAI** y **Verifactu**.

    **SIGN ES** proporciona un entorno de prueba para fines de integración y pruebas, así como un entorno de producción dedicado para realizar el envío de transacciones oficiales (facturas). Para acceder al entorno de producción, [contacta con nuestro equipo de ventas](mailto:sales@fiskaly.com).

...

    - tiene URLs orientadas a recursos,
...
    - usa verbos y códigos de estado HTTP estándar, y  
...

...

...

...

...

...

...
paths:
...
  '/exports/{export_id}':
    put:
      summary: Crear exportación
      description: Este endpoint crea una exportación asíncrona de facturas para los criterios aplicados.
      tags:
        - 'Exportar datos'
      security:
        - JWT: []
      operationId: createExport
      parameters:
        - $ref: '#/components/parameters/ExportIdentifier'
        - $ref: '#/components/parameters/InvoiceSeriesQuery'
        - $ref: '#/components/parameters/InvoiceIssuedAtFromQuery'
        - $ref: '#/components/parameters/InvoiceIssuedAtToQuery'
        - $ref: '#/components/parameters/ClientIdentifierQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                metadata:
                  $ref: '#/components/schemas/Metadata'
              additionalProperties: false
      responses:
        '200':
          $ref: '#/components/responses/ExportResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '409':
          $ref: '#/components/responses/ResourceConflictResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    patch:
      summary: Actualizar exportación
      description: Este endpoint permite actualizar los metadatos de una exportación.
      tags:
        - 'Exportar datos'
      security:
        - JWT: []
      operationId: updateExport
      parameters:
        - $ref: '#/components/parameters/ExportIdentifier'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                metadata:
                  $ref: '#/components/schemas/Metadata'
              additionalProperties: false
      responses:
        '200':
          $ref: '#/components/responses/ExportResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '404':
          $ref: '#/components/responses/ResourceNotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    get:
      summary: Obtener una exportación
      description: Este endpoint obtiene un recurso que proporciona el estado de la solicitud de exportación.
      tags:
        - 'Exportar datos'
      security:
        - JWT: []
      operationId: retrieveExport
      parameters:
        - $ref: '#/components/parameters/ExportIdentifier'
      responses:
        '200':
          $ref: '#/components/responses/ExportResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '404':
          $ref: '#/components/responses/ResourceNotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /exports:
    get:
      summary: Listar exportaciones
      description: Este endpoint obtiene una lista de exportaciones solicitadas del contribuyente.
      tags:
        - 'Exportar datos'
      security:
        - JWT: []
      operationId: listExports
      parameters:
        - $ref: '#/components/parameters/PaginationLimitQuery'
        - $ref: '#/components/parameters/PaginationTokenQuery'
      responses:
        '200':
          $ref: '#/components/responses/ExportsResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  '/exports/{export_id}.zip':
    get:
      summary: Descargar exportación
      description: Este endpoint obtiene el archivo zip de una exportación.
      tags:
        - 'Exportar datos'
      security:
        - JWT: []
      operationId: downloadExport
      parameters:
        - $ref: '#/components/parameters/ExportIdentifier'
      responses:
        '200':
          $ref: '#/components/responses/ExportZipResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '404':
          $ref: '#/components/responses/ResourceNotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
components:
  parameters:
...
    ExportIdentifier:
      description: 'Identifica una exportación con un identificador aleatorio en formato [UUIDv4](#section/Producto/Recursos).'
      in: path
      name: export_id
      required: true
      schema:
        $ref: '#/components/schemas/UniversallyUniqueIdentifierV4'
...
    InvoiceSeriesQuery:
      description: Identifica el número de serie de una factura en un determinado año.
      in: query
      name: series
      required: false
      schema:
        $ref: '#/components/schemas/TicketbaiInvoiceIdentifier'
    InvoiceIssuedAtFromQuery:
      description: Define la fecha más temprana posible de la factura (inclusive).
      in: query
      name: issued_at_from
      required: false
      schema:
        $ref: '#/components/schemas/TimestampFormat19'
    InvoiceIssuedAtToQuery:
      description: Define la última fecha posible de la factura (exclusive).
      in: query
      name: issued_at_to
      required: false
      schema:
        $ref: '#/components/schemas/TimestampFormat19'
    ClientIdentifierQuery:
      description: Define el identificador del cliente que emite una factura.
      in: query
      name: client_id
      required: false
      schema:
        $ref: '#/components/schemas/UniversallyUniqueIdentifierV4'
    PaginationLimitQuery:
      in: query
      name: limit
      required: false
      schema:
        $ref: '#/components/schemas/PaginationLimit'
    PaginationTokenQuery:
      in: query
      name: token
      required: false
      schema:
        $ref: '#/components/schemas/PaginationToken'
  responses:
...
    ExportsResponse:
      description: Devuelve una lista de exportaciones.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExportResources'
    ExportResponse:
      description: Devuelve una exportación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExportResource'
    ExportZipResponse:
      description: Returns an export zip archive.
      content:
        application/zip:
          schema:
            $ref: '#/components/schemas/Binary'
...
    BadRequestResponse:
      description: Solicitud incorrecta (no válida)
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
    UnauthorizedAccessResponse:
      description: Acceso no autorizado.
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
    ResourceNotFoundResponse:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
    ResourceConflictResponse:
      description: El recurso solicitado tiene un conflicto.
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
...
    InternalServerErrorResponse:
      description: Ocurrió un error interno del servidor
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
...
  schemas:
    ErrorResponse:
      description: Representación de respuestas de error.
      type: object
      required:
        - code
        - message
        - status
        - error
      properties:
        status:
          $ref: '#/components/schemas/ErrorStatus'
        code:
          $ref: '#/components/schemas/ErrorCode'
        error:
          $ref: '#/components/schemas/ErrorText'
        message:
          $ref: '#/components/schemas/ErrorMessage'
      additionalProperties: false
    ErrorStatus:
      description: Código de estado de error HTTP basado en números enteros.
      type: integer
      format: int32
      example: 401
    ErrorCode:
      description: Representación de un nombre de error único del código de estado de error HTTP.
      type: string
      example: E_UNAUTHORIZED_ACCESS
    ErrorText:
      description: Representación de una breve descripción del código de estado de error HTTP.
      type: string
      example: Unauthorized Access
    ErrorMessage:
      description: Representación de una descripción larga del código de estado de error HTTP.
      type: string
      example: Not authorized to perform this operation.
...
    ExportResources:
      description: Representación de recursos de exportación.
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/ExportResource'
        pagination:
          $ref: '#/components/schemas/Pagination'
      additionalProperties: false
    ExportResource:
      description: Representación de un recurso de exportación.
      type: object
      required:
        - content
      properties:
        content:
          $ref: '#/components/schemas/Export'
        metadata:
          $ref: '#/components/schemas/Metadata'
      additionalProperties: false
...
    Export:
      description: Representación de una exportación de factura solicitada.
      type: object
      required:
        - id
        - state
        - requested_at
      properties:
        id:
          $ref: '#/components/schemas/UniversallyUniqueIdentifierV4'
        state:
          $ref: '#/components/schemas/ExportState'
        requested_at:
          $ref: '#/components/schemas/ExportRequestedTimestamp'
        started_at:
          $ref: '#/components/schemas/ExportStartedTimestamp'
        finished_at:
          $ref: '#/components/schemas/ExportFinishedTimestamp'
        expires_at:
          $ref: '#/components/schemas/ExportExpiresTimestamp'
        download:
          $ref: '#/components/schemas/ExportDownload'
      additionalProperties: false
    ExportRequestedTimestamp:
      description: 'La fecha y hora (en zona horaria UTC y formato `DD-MM-AAAA hh:mm:ss`) de cuando la exportación ha sido solicitada.'
      allOf:
        - $ref: '#/components/schemas/TimestampFormat19'
    ExportStartedTimestamp:
      description: 'La fecha y hora (en zona horaria UTC y formato `DD-MM-AAAA hh:mm:ss`) de cuando la exportación ha comenzado.'
      allOf:
        - $ref: '#/components/schemas/TimestampFormat19'
    ExportFinishedTimestamp:
      description: 'La fecha y hora (en zona horaria UTC y formato `DD-MM-AAAA hh:mm:ss`) de cuando ha finalizado la exportación.'
      allOf:
        - $ref: '#/components/schemas/TimestampFormat19'
    ExportExpiresTimestamp:
      description: 'La fecha y hora (en zona horaria UTC y formato `DD-MM-AAAA hh:mm:ss`) de cuando caduca la exportación.'
      allOf:
        - $ref: '#/components/schemas/TimestampFormat19'
    ExportDownload:
      description: Información sobre una descarga de exportación.
      type: object
      required:
        - url
        - size
        - checksum
      properties:
        url:
          $ref: '#/components/schemas/UniformResourceIdentifier'
        size:
          $ref: '#/components/schemas/ByteSize'
        checksum:
          $ref: '#/components/schemas/ExportChecksum'
      additionalProperties: false
    ExportChecksum:
      description: Representa la información del checksum de una exportación descargable.
      discriminator:
        propertyName: type
        mapping:
          MD5: '#/components/schemas/MessageDigestMethod5'
      oneOf:
        - $ref: '#/components/schemas/MessageDigestMethod5'
    MessageDigestMethod5:
      description: Información de un checksum MD5.
      type: object
      required:
        - type
        - value
      properties:
        type:
          type: string
          enum:
            - MD5
        hash:
          description: Suma de verificación MD5 de una carga útil.
          type: string
          pattern: '^[0-9a-fA-F]{32}$'
          minLength: 32
          maxLength: 32
          example: c6779ec2960296ed9a04f08d67f64422
      additionalProperties: false
...
    TicketbaiInvoiceIdentifier:
      description: Representación del esquema del identificador predefinido según la normativa española.
      type: string
      pattern: '^[0-9A-Z_/\-\.]{1,20}$'
      minLength: 1
      maxLength: 20
      example: '2022'
...
    TimestampFormat19:
      description: 'Timestamp en `DD-MM-AAAA hh:mm:ss`.'
      type: string
      minLength: 19
      maxLength: 19
      example: '24-12-1992 15:35:01'
...
    ExportState:
      description: 'Representa el estado de la exportación. Por defecto, una exportación se crea en estado `REQUESTED` (SOLICITADA). En cuanto comienza la exportación, tendrá el estado `STARTED` (INICIADA). Una vez ha finalizado la exportación, se indica el estado `FINISHED`  (FINALIZADA) y está disponible para descarga. Cuando una exportación ha caducado, se encuentra en estado `EXPIRED`  (CADUCADA) y ya no podrá ser descargada. Si una exportación falla, por cualquier motivo, resulta en el estado `FAILED` (FALLIDA)'
      type: string
      enum:
        - REQUESTED
        - STARTED
        - FINISHED
        - EXPIRED
        - FAILED
...
    UniversallyUniqueIdentifierV4:
      description: |-
        Un identificador único universal versión 4 (UUIDv4), es una etiqueta empleada para identificar objetos de forma única en
        sistemas informáticos. La representación de UUIDs está definida dentro del [RFC 4122](https://www.rfc-editor.org/rfc/rfc4122).
      type: string
      format: uuid
      pattern: '[a-f0-9]{8}-?[a-f0-9]{4}-?4[a-f0-9]{3}-?[89ab][a-f0-9]{3}-?[a-f0-9]{12}'
      example: 1c81cb86-c2e8-4074-afc3-a0601b2bf063
      x-go-type: uuid.UUID
      x-go-type-import:
        name: uuid
        path: github.com/google/uuid
...
    Metadata:
      description: |-
        Puede emplear este parámetro para adjuntar datos en formato clave/valor al objeto.  
        Los metadatos son útiles para almacenar datos estructurados que complementan el objeto.

        Se pueden adjuntar metadatos en los métodos de actualización de cada recurso. Si se incluye una clave con un valor vacío, se eliminará esa pareja clave/valor.

        Si se adjuntan metadatos a un objeto, estos serán devueltos en las respuestas a peticiones de ese objeto.

        **Nota**: Se pueden especificar hasta 20 claves, cada clave puede contener hasta 40 caracteres, y los valores pueden contener hasta 500 caracteres.
      type: object
      maxProperties: 20
      additionalProperties:
        type: string
        maxLength: 500
      example:
        my_property_1: '1234'
        my_property_2: 'https://my-internal-system/path/to/resource/1234'
...
    Binary:
      description: Representación de un archivo binario.
      type: string
      format: binary
...
    Pagination:
      description: 'Representa los datos de paginación de un endpoint de listado. Si no se especifican, los resultados actuales son la última página.'
      type: object
      required:
        - next
        - token
        - limit
      properties:
        next:
          $ref: '#/components/schemas/PaginationNext'
        token:
          $ref: '#/components/schemas/PaginationToken'
        limit:
          $ref: '#/components/schemas/PaginationLimit'
    PaginationNext:
      description: 'Representa la URL de petition la siguiente página de resultados. La petición está compuesta por el recurso de listado, el token de paginación, y el límite de resultados a presentar.'
      allOf:
        - $ref: '#/components/schemas/UniformResourceIdentifier'
    PaginationLimit:
      description: Representa el límite de resultados por página para un endpoint de listado.
      type: integer
      minimum: 1
      maximum: 100
      default: 10
    PaginationToken:
      description: 'Representa el token de paginación, que se emplea como puntero hacia la siguiente página de resultados.'
      type: string
      pattern: '^(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{4}|[A-Za-z0-9+/]{3}=|[A-Za-z0-9+/]{2}={2})$'
      maxLength: 1024
    UniformResourceIdentifier:
      description: Representa un identificador de recursos uniforme.
      type: string
    ByteSize:
      description: El tamaño en bytes de la carga útil.
      type: integer
      example: 1234
  securitySchemes:
    JWT:
      description: |-
        JSON Web Token (JWT) empleado para control de acceso y autorización.
        - Ejemplo de uso: `Bearer <JWT>`
        - Ejemplo de encabezado HTTP: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
      type: http
      scheme: bearer
      bearerFormat: JWT